변수명 및 파일명 변경, 파일구조 변경
