+++ app/src/view/main/chat/ChatMain.js
@@ -3,12 +3,13 @@ import {inject, observer} from 'mobx-react';
 import Colors from 'react-native/Libraries/NewAppScreen/components/Colors';
 import {StyleSheet, Text, SafeAreaView, View, ScrollView} from 'react-native';
 import ChatMainHeader from './ChatMainHeader';
-import {Main} from "./list/chatListMain";
+import {ChatListMain} from "./list/ChatListMain";
+import {ChatListHeader} from "./list/ChatListHeader";
 
 class ChatMain extends Component {
     // 컴포넌트가 만들어지고 첫 렌더링을 다 마친 후 실행되는 메소드입니다.
     // 이 안에서 다른 JavaScript 프레임워크를 연동하거나,
-    // setTimeout, setInterval 및 AJAX 처리 등을 넣습니다.
+    // setTimeout, setInterval 및 AJAX 처리 등을 넣습니다.s
     componentDidMount() {
     }
 
@@ -25,8 +26,9 @@ class ChatMain extends Component {
         return (
             <SafeAreaView style={styles.body}>
                 {/*<ChatMainHeader/>*/}
-                <Main />
-                <ScrollView style={styles.scrollView}/>
+                <ChatListHeader/>
+                <ChatListMain />
+                {/*<ScrollView style={styles.scrollView}/>*/}
             </SafeAreaView>
         );
     }
