--- app/src/view/main/chat/list/detail.js
@@ -1,9 +1,0 @@
-import * as React from 'react';
-import {View, Text} from 'react-native';
-function DetailsScreen() {
-  return (
-    <View style={{flex: 1, alignItems: 'center', justifyContent: 'center'}}>
-      <Text>Details Screen</Text>
-    </View>
-  );
-}
